refactor(dapui, dui) rename ui to dui
